Add in base ValidatorTester class.
